[spi] simplify DataTable interface
