+++ thirdeye-spi/src/main/java/ai/startree/thirdeye/spi/detection/v2/DataTable.java
@@ -14,31 +14,11 @@
 package ai.startree.thirdeye.spi.detection.v2;
 
 import ai.startree.thirdeye.spi.dataframe.DataFrame;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 // todo cyril simplify interface - most of the methods are not used
 public interface DataTable extends OperatorResult {
-
-  static Object[] getRow(final DataTable dataTable, final int rowIdx) {
-    int columnCount = dataTable.getColumnCount();
-    Object[] row = new Object[columnCount];
-    for (int colIdx = 0; colIdx < columnCount; colIdx++) {
-      row[colIdx] = dataTable.getObject(rowIdx, colIdx);
-    }
-    return row;
-  }
-
-  static Map<String, Object> getRecord(final List<String> columnNames, final Object[] event) {
-    Map<String, Object> record = new HashMap<>();
-    for (int i = 0; i < columnNames.size(); i++) {
-      record.put(columnNames.get(i), event[i]);
-    }
-    return record;
-  }
-
-  int getRowCount();
 
   int getColumnCount();
